fix for vscode/libreoffice

+++ capnp/write.capnp
@@ -8,7 +8,7 @@ struct Write {
   fi    @4 :FuseFileInfo;
 
   struct FuseFileInfo {
-    flags         @0  :Int8;
+    flags         @0  :Int64;
     writepage     @1  :UInt64;
     directIo      @2  :UInt64;
     keepCache     @3  :UInt64;
